requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
pycparser==2.22
PyJWT==2.8.0
pytest==8.3.2
pytest-mock==3.14.0
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
scapy==2.5.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
